--- src/pages/BoardPage.jsx
@@ -1,12 +1,0 @@
-import Board from '../components/boardPage/Board/Board';
-import AppLayout from '../components/common/AppLayout';
-
-const BoardPage = () => {
-  return (
-    <AppLayout>
-      <Board />
-    </AppLayout>
-  );
-};
-
-export default BoardPage;
