+++ src/pages/PostPage.jsx
@@ -1,5 +1,5 @@
+import Post from '../components/boardPage/Post/Post';
 import AppLayout from '../components/common/AppLayout';
-import Post from '../components/boardPage/PostEditor/Post';
 
 const PostPage = () => {
   return (
